refactor(chips): Deprecate md-on-append in favor of others.
The usage of `md-on-append` was not well documented and it's
behavior was inconsistent and confusing; many users assumed
it was a simple notification of chip additions which caused
issues.

After much discussion, we have renamed `md-on-append` to
`md-transform-chip` and provided a new `md-on-add` method
that is strictly a notification.

Additionally, we have updated the docs and functionality of
`md-transform-chip` to show expected return values and their
associated behavior.

This new behavior also adds support for simlultaneously using
an autocomplete to select an existing value along with the
ability to create new chips. The most common case for this is
a tag system which shows existing tags, but also allows you to
create new ones.

Demos have been updated to show new functionality as well as
to workaround a few display issues with the contact chips demo
(#4450).

_**Note:** This work supercedes PR #3816 which can be closed when
this is merged._

BREAKING CHANGE:
`md-on-append` has been renamed/deprecated in favor of
`md-transform-chip` or the simple notifier `md-on-add`.

We expect to remove this completely in 1.0, so please update
your code to use one of the new methods.

Fixes #4666. Fixes #4193. Fixes #4412. Fixes #4863.